images_train/requirements.txt
torch>=2.0.0
transformers>=4.30.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
easyocr>=1.7.0
reportlab>=4.0.0